gnu: libaio: Skip tests which fail on btrfs
Test cases #18 and #21 of libaio depend on the behavior of O_DIRECT on
the filesystem. However, on Btrfs, although O_DIRECT is nominally
reported as supported by the filesystem, the behavior can not be relied
on when checksumming is enabled on the filesystem. See
https://patchwork.kernel.org/project/linux-btrfs/patch/5d52220b-177d-72d4-7825-dbe6cbf8722f@inwind.it/

Skip these two test cases unconditionally in libaio, since they do not
pass when run on a default Btrfs filesystem.

* gnu/local.mk (patches): Include patch file
* gnu/packages/patches/libaio-btrfs-test.patch: Patch disabling tests
* gnu/packages/linux.scm (libaio): Include patch